Update route tables in the Coinbase VPC
Create two route tables, one for public subnet traffic and another for private. The private route table routes all outbound traffic to the NAT Gateway in the public subnet while the public subnet routes all outbound traffic to the Internet Gateway. The public subnet is only with subnet_a, which is the public subnet.